Fix -Wnarrowing in LinuxHardwareI2C::requestFrom
i2c_msg.len is __u16. requestedBytes (int) and count (size_t) are wider
types; assign them with explicit (__u16) casts to silence the narrowing
conversion warnings.

+++ cores/ardulinux/linux/LinuxHardwareI2C.cpp
@@ -145,13 +145,13 @@ namespace arduino {
                 {  // Start address
                 .addr = address,
                 .flags = 0, // write
-                .len = requestedBytes,
+                .len = (__u16)requestedBytes,
                 .buf = (unsigned char *) TXbuf
                 },
                 { // Read buffer
                 .addr = address,
                 .flags = I2C_M_RD, // read
-                .len = count,
+                .len = (__u16)count,
                 .buf = (unsigned char *)RXbuf
                 }
             };
